ROS Drivers for OVC5 (#46)
* ROS1 driver for ovc


* libovc


* libovc based ros drivers


* Default app uses libovc


* Draw image wrt screen width


* Fixed color issue and added makefile for testing convenience


* Remove iostream


* Dynamic config document


* Dynamic reconfigure for ROS1


* Get ethernet packet definitions from libovc


* Publish -> Send && Publisher -> Client


* Adding rx packet type to device driver


* Clear up compiler warnings


* Socket read call fix


* Fix compiler warnings and do some formatting along the way


* Subscriber -> Server


* Dynamic reconfigure plumbing


* Symlink packet header file instead of messy dependency


* Print on device when receiving packets


* Fix locking issue


* ROS2 add params


* Formatting with clang-format to match style


* Correct code format

+++ ovc5/software/camera_device_driver/src/i2c_driver.cpp
@@ -52,7 +52,7 @@ int I2CDriver::writeRegister(const regop_t& regop)
   // We need to add one byte for address
   std::vector<uint8_t> payload(reg_size + 1);
   payload[0] = regop.addr & 0xFF;
-  for (size_t i = 1; i <= reg_size; ++i)
+  for (int i = 1; i <= reg_size; ++i)
     payload[i] = regop.i32 >> (8 * (reg_size - i));
   return i2c_smbus_write_i2c_block_data(
       i2c_fd, regop.addr >> 8, payload.size(), &payload[0]);
